Restructure interpretation modules to provide default implementations
