+++ src/Language/Syntactic/Interpretation/Evaluation.hs
@@ -1,20 +1,34 @@
+{-# LANGUAGE DefaultSignatures #-}
+
 module Language.Syntactic.Interpretation.Evaluation where
 
 
 
 import Language.Syntactic.Syntax
+import Language.Syntactic.Interpretation.Semantics
 
 
-
--- | The denotation of a symbol with the given signature
-type family   Denotation sig
-type instance Denotation (Full a)    = a
-type instance Denotation (a :-> sig) = a -> Denotation sig
 
 class Eval expr
   where
     -- | Evaluation of expressions
     evaluate :: expr a -> Denotation a
+
+    default evaluate :: Semantic expr => expr a -> Denotation a
+    evaluate = evaluateDefault
+    {-# INLINABLE evaluate #-}
+
+-- | Default implementation of 'evaluate'
+evaluateDefault :: Semantic expr => expr a -> Denotation a
+evaluateDefault = evaluate . semantics
+{-# INLINE evaluateDefault #-}
+
+instance Eval Semantics
+  where
+    {-# SPECIALIZE instance Eval Semantics #-}
+    {-# INLINABLE evaluate #-}
+    evaluate (Sem _ a) = a
+
 
 instance Eval dom => Eval (AST dom)
   where
